port entire project from .NET Framework 4.7.2 to .NET 8.0; correct issue with KFDtool.Shared.Utility.Decompress failing to decompress the entire stream and returning a partial decompression, instead of using a loop simnply use .NET 8's CopyTo mechanism to copy the full stream; refactor and replace the entire EKC container library, this uses the one from my port of the KFDtool UI (KFDEKC) which includes file support for UKEK and LLA to be included in EKC files (I did not port the UI work for this because my KFDEKC fork of the UI has extensive UI changes making this difficult to accomplish);
