+++ sw/control/KFDEKC.Container/FileStructure/DKF/DKFInnerValueData.cs
@@ -1,0 +1,21 @@
+﻿// SPDX-License-Identifier: MIT
+/*
+* KFDtool Container (EKC) Editor
+* MIT Open Source. Use is subject to license terms.
+* DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
+*
+*/
+
+using System.Xml.Serialization;
+
+namespace KFDEKC.Container.FileStructure.DKF
+{
+    /// <summary>
+    /// 
+    /// </summary>
+    [XmlRoot("valuedata")]
+    public class DKFInnerValueData
+    {
+        /* stub */
+    }
+}
