+++ sw/control/KFDEKC.Container/FileStructure/DKF/DKFOuterPersonality.cs
@@ -1,34 +1,81 @@
-﻿using System.Xml.Serialization;
+﻿// SPDX-License-Identifier: MIT
+/*
+* KFDtool Container (EKC) Editor
+* MIT Open Source. Use is subject to license terms.
+* DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
+*
+*/
 
-namespace KFDtool.Container
+using System.Xml.Serialization;
+
+namespace KFDEKC.Container.FileStructure.DKF
 {
+    /// <summary>
+    /// 
+    /// </summary>
     [XmlRoot("personality")]
     public class DKFOuterPersonality
     {
+        /// <summary>
+        /// 
+        /// </summary>
         [XmlAttribute("pdlversion")]
         public string PDLVersion { get; set; }
 
+        /// <summary>
+        /// 
+        /// </summary>
         [XmlAttribute("defaultversion")]
         public string DefaultVersion { get; set; }
 
+        /// <summary>
+        /// 
+        /// </summary>
         [XmlAttribute("valueversion")]
         public string ValueVersion { get; set; }
 
+        /// <summary>
+        /// 
+        /// </summary>
         [XmlAttribute("devicetype")]
         public string DeviceType { get; set; }
 
+        /// <summary>
+        /// 
+        /// </summary>
         [XmlAttribute("targetversionmajor")]
         public string TargetVersionMajor { get; set; }
 
+        /// <summary>
+        /// 
+        /// </summary>
         [XmlAttribute("targetversionminor")]
         public string TargetVersionMinor { get; set; }
 
+        /// <summary>
+        /// 
+        /// </summary>
         [XmlAttribute("type")]
         public string Type { get; set; }
 
+        /// <summary>
+        /// 
+        /// </summary>
         [XmlAttribute("percategory")]
         public string PerCategory { get; set; }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        public string ParamData { get; set; }
+
+        /*
+        ** Methods
+        */
+
+        /// <summary>
+        /// 
+        /// </summary>
         public DKFOuterPersonality()
         {
             PDLVersion = "2";
@@ -40,7 +87,5 @@ namespace KFDtool.Container
             Type = "User";
             PerCategory = "Encrypted";
         }
-
-        public string ParamData { get; set; }
     }
 }
