+++ sw/control/KFDEKC.Container/FileStructure/EKC/GroupItem.cs
@@ -1,0 +1,79 @@
+﻿// SPDX-License-Identifier: MIT
+/*
+* KFDtool Container (EKC) Editor
+* MIT Open Source. Use is subject to license terms.
+* DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
+*
+*/
+
+using System.Collections.Generic;
+using System.ComponentModel;
+
+namespace KFDEKC.Container.FileStructure.EKC
+{
+    /// <summary>
+    /// 
+    /// </summary>
+    public class GroupItem : INotifyPropertyChanged
+    {
+        private string name;
+
+        /// <summary>
+        /// 
+        /// </summary>
+        public int Id { get; set; }
+
+        /// <summary>
+        /// 
+        /// </summary>
+        public string Name
+        {
+            get { return name; }
+            set
+            {
+                if (name != value)
+                {
+                    name = value;
+                    NotifyPropertyChanged("Name");
+                }
+            }
+        }
+
+        /// <summary>
+        /// 
+        /// </summary>
+        public List<int> Keys { get; set; }
+
+        /*
+        ** Events
+        */
+
+        /// <summary>
+        /// 
+        /// </summary>
+        public event PropertyChangedEventHandler PropertyChanged;
+
+        /*
+        ** Methods
+        */
+
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="propName"></param>
+        public void NotifyPropertyChanged(string propName)
+        {
+            if (PropertyChanged != null)
+                PropertyChanged(this, new PropertyChangedEventArgs(propName));
+        }
+
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <returns></returns>
+        public override string ToString()
+        {
+            return Name.ToString();
+        }
+    }
+}
