+++ sw/control/KFDEKC.Container/FileStructure/EKC/RSIKeyItem.cs
@@ -1,0 +1,23 @@
+﻿// SPDX-License-Identifier: MIT
+/*
+* KFDtool Container (EKC) Editor
+* MIT Open Source. Use is subject to license terms.
+* DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
+*
+*/
+
+using System.ComponentModel;
+
+namespace KFDEKC.Container.FileStructure.EKC
+{
+    /// <summary>
+    /// 
+    /// </summary>
+    public class RSIKeyItem : KeyItem
+    {
+        /// <summary>
+        /// 
+        /// </summary>
+        public int RsiId { get; set; }
+    }
+}
