--- sw/control/KFDtool.Container/DKFInnerValueData.cs
@@ -1,10 +1,0 @@
-﻿using System.Xml.Serialization;
-
-namespace KFDtool.Container
-{
-    [XmlRoot("valuedata")]
-    public class DKFInnerValueData
-    {
-        /* stub */
-    }
-}
