--- sw/control/KFDtool.Container/DKFOuterParamData.cs
@@ -1,10 +1,0 @@
-﻿using System.Xml.Serialization;
-
-namespace KFDtool.Container
-{
-    [XmlRoot("paramdata")]
-    public class DKFOuterParamData
-    {
-        /* stub */
-    }
-}
