--- sw/control/KFDtool.Container/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="NLog" version="5.2.5" targetFramework="net472" />
-</packages>
