+++ sw/control/KFDtool.Gui/Settings.cs
@@ -1,4 +1,4 @@
-﻿using KFDtool.Container;
+﻿using KFDEKC.Container.FileStructure.EKC;
 using KFDtool.P25.TransferConstructs;
 using System.Diagnostics;
 using System.Reflection;
