+++ sw/control/KFDtool.sln
@@ -17,9 +17,9 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "KFDtool.Shared", "KFDtool.S
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HidLibrary", "HidLibrary\HidLibrary.csproj", "{12C7DCF8-287A-4CE9-A04D-ED8E037CD6AA}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "KFDtool.Container", "KFDtool.Container\KFDtool.Container.csproj", "{4E99EC83-6EE7-49E0-A9AF-6674A53F578C}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "KFDtool.Elite", "KFDtool.Elite\KFDtool.Elite.csproj", "{6C166F22-1FC0-4C52-A2ED-5A46CF9216BD}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "KFDEKC.Container", "KFDEKC.Container\KFDEKC.Container.csproj", "{76C43B39-6CCD-0F34-1F8F-D0E0E6C1E4F5}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -115,18 +115,6 @@ Global
 		{12C7DCF8-287A-4CE9-A04D-ED8E037CD6AA}.Release|x64.Build.0 = Release|x64
 		{12C7DCF8-287A-4CE9-A04D-ED8E037CD6AA}.Release|x86.ActiveCfg = Release|x86
 		{12C7DCF8-287A-4CE9-A04D-ED8E037CD6AA}.Release|x86.Build.0 = Release|x86
-		{4E99EC83-6EE7-49E0-A9AF-6674A53F578C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4E99EC83-6EE7-49E0-A9AF-6674A53F578C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4E99EC83-6EE7-49E0-A9AF-6674A53F578C}.Debug|x64.ActiveCfg = Debug|x64
-		{4E99EC83-6EE7-49E0-A9AF-6674A53F578C}.Debug|x64.Build.0 = Debug|x64
-		{4E99EC83-6EE7-49E0-A9AF-6674A53F578C}.Debug|x86.ActiveCfg = Debug|x86
-		{4E99EC83-6EE7-49E0-A9AF-6674A53F578C}.Debug|x86.Build.0 = Debug|x86
-		{4E99EC83-6EE7-49E0-A9AF-6674A53F578C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4E99EC83-6EE7-49E0-A9AF-6674A53F578C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{4E99EC83-6EE7-49E0-A9AF-6674A53F578C}.Release|x64.ActiveCfg = Release|x64
-		{4E99EC83-6EE7-49E0-A9AF-6674A53F578C}.Release|x64.Build.0 = Release|x64
-		{4E99EC83-6EE7-49E0-A9AF-6674A53F578C}.Release|x86.ActiveCfg = Release|x86
-		{4E99EC83-6EE7-49E0-A9AF-6674A53F578C}.Release|x86.Build.0 = Release|x86
 		{6C166F22-1FC0-4C52-A2ED-5A46CF9216BD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{6C166F22-1FC0-4C52-A2ED-5A46CF9216BD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6C166F22-1FC0-4C52-A2ED-5A46CF9216BD}.Debug|x64.ActiveCfg = Debug|x64
@@ -139,6 +127,18 @@ Global
 		{6C166F22-1FC0-4C52-A2ED-5A46CF9216BD}.Release|x64.Build.0 = Release|x64
 		{6C166F22-1FC0-4C52-A2ED-5A46CF9216BD}.Release|x86.ActiveCfg = Release|x86
 		{6C166F22-1FC0-4C52-A2ED-5A46CF9216BD}.Release|x86.Build.0 = Release|x86
+		{76C43B39-6CCD-0F34-1F8F-D0E0E6C1E4F5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{76C43B39-6CCD-0F34-1F8F-D0E0E6C1E4F5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{76C43B39-6CCD-0F34-1F8F-D0E0E6C1E4F5}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{76C43B39-6CCD-0F34-1F8F-D0E0E6C1E4F5}.Debug|x64.Build.0 = Debug|Any CPU
+		{76C43B39-6CCD-0F34-1F8F-D0E0E6C1E4F5}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{76C43B39-6CCD-0F34-1F8F-D0E0E6C1E4F5}.Debug|x86.Build.0 = Debug|Any CPU
+		{76C43B39-6CCD-0F34-1F8F-D0E0E6C1E4F5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{76C43B39-6CCD-0F34-1F8F-D0E0E6C1E4F5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{76C43B39-6CCD-0F34-1F8F-D0E0E6C1E4F5}.Release|x64.ActiveCfg = Release|Any CPU
+		{76C43B39-6CCD-0F34-1F8F-D0E0E6C1E4F5}.Release|x64.Build.0 = Release|Any CPU
+		{76C43B39-6CCD-0F34-1F8F-D0E0E6C1E4F5}.Release|x86.ActiveCfg = Release|Any CPU
+		{76C43B39-6CCD-0F34-1F8F-D0E0E6C1E4F5}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
